contracts-bedrock: interop dep set refactor
Proof of concept refactor on what the dependency set management could
look like.

The problem is that there is onchain management of the dependency set
and the values are kept consistent between L1 and each L2. It is more
simple to keep a single source of truth on L1. We will need to move the
logic that was previously onchain L2 to the block builder for this to
work.

There need to be updates done to the `L2ToL2CrossDomainMessenger`
as well as the `CrossL2Inbox` for this to be complete, just remove
the concept of dep set completely.